third_party/WebKit/LayoutTests/external/wpt/media-source/mediasource-addsourcebuffer-mode-expected.txt

This is a testharness.js-based test.
PASS addSourceBuffer() sets SourceBuffer.mode to 'segments' when the generate timestamps flag is false
FAIL addSourceBuffer() sets SourceBuffer.mode to 'sequence' when the generate timestamps flag is true assert_unreached: Browser does not support the audio/aac and audio/mpeg MIME types used in this test Reached unreachable code
Harness: the test ran to completion.
